Separate AdminWeb from the cb-network controller
- Copy controller.go to admin-web.go
- Remove admin-web source code from controller.go
- Remove controller source code from admin-web.go
- Update gitignore to ignore admin-web, the execution file
