+++ .gitignore
@@ -2,6 +2,7 @@
 *.exe
 poc-cb-net/cmd/agent/agent
 poc-cb-net/cmd/controller/controller
+poc-cb-net/cmd/admin-web/admin-web
 
 # Ignore configuration directory
 # to avoid leaking sensitive information and frequent update
